requirements.txt
aiogram
environs
python-dotenv
asincpg
